preparations for hw sobel filter

+++ src/application/hdmi_input.c
@@ -123,7 +123,7 @@ LOOP:
             }
 
 
-            MBOX_PUT(hdmi_input->mb, buffers[buf.index].start);
+            MBOX_PUT(hdmi_input->mb,(uint32_t)(buffers[buf.index].start));
 	
             
             if (-1 == xioctl(hdmi_input->fd, VIDIOC_QBUF, &buf))
